// web/src/app/enums/keycode.ts
export enum Keycode {
    CONTROL = 'Keycode.CONTROL',
    SHIFT = 'Keycode.SHIFT',
    ALT = 'Keycode.ALT',
    WINDOWS = 'Keycode.WINDOWS',
    OPTION = 'Keycode.OPTION',
    COMMAND = 'Keycode.COMMAND',
    COMMA = 'Keycode.COMMA',
    PERIOD = 'Keycode.PERIOD',
    FORWARD_SLASH = 'Keycode.FORWARD_SLASH',
    SEMI_COLON = 'Keycode.SEMI_COLON',
    QUOTE = 'Keycode.QUOTE',
    LEFT_BRACKET = 'Keycode.LEFT_BRACKET',
    RIGHT_BRACKET = 'Keycode.RIGHT_BRACKET',
    BACKSLASH = 'Keycode.BACKSLASH',
    GRAVE_ACCENT = 'Keycode.GRAVE_ACCENT',
    MINUS = 'Keycode.MINUS',
    EQUALS = 'Keycode.EQUALS',
    ARROW_UP = 'Keycode.UP_ARROW',
    ARROW_DOWN = 'Keycode.DOWN_ARROW',
    ARROW_LEFT = 'Keycode.LEFT_ARROW',
    ARROW_RIGHT = 'Keycode.RIGHT_ARROW',
    CAPS_LOCK = 'Keycode.CAPS_LOCK',
    ENTER = 'Keycode.ENTER',
    TAB = 'Keycode.TAB',
    SPACE = 'Keycode.SPACE',
    BACKSPACE = 'Keycode.BACKSPACE',
    ESCAPE = 'Keycode.ESCAPE',
    F1 = 'Keycode.F1',
    F2 = 'Keycode.F2',
    F3 = 'Keycode.F3',
    F4 = 'Keycode.F4',
    F5 = 'Keycode.F5',
    F6 = 'Keycode.F6',
    F7 = 'Keycode.F7',
    F8 = 'Keycode.F8',
    F9 = 'Keycode.F9',
    F10 = 'Keycode.F10',
    F11 = 'Keycode.F11',
    F12 = 'Keycode.F12',
    PRINT_SCREEN = 'Keycode.PRINT_SCREEN',
    SCROLL_LOCK = 'Keycode.SCROLL_LOCK',
    PAUSE = 'Keycode.PAUSE',
    INSERT = 'Keycode.INSERT',
    HOME = 'Keycode.HOME',
    END = 'Keycode.END',
    PAGE_UP = 'Keycode.PAGE_UP',
    PAGE_DOWN = 'Keycode.PAGE_DOWN',
    DELETE = 'Keycode.DELETE',
    NUM_LOCK = 'Keycode.KEYPAD_NUMLOCK', // TODO = More keypad buttons
    POWER = 'Keycode.POWER'
}